requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
requests==2.32.4
pydantic==2.11.7
pypdf==5.9.0
pinecone==7.3.0
sentence-transformers==2.2.2
scipy==1.10.1
numpy==1.24.4
starlette==0.47.2
python-multipart==0.0.20
